requirements.txt
customtkinter>=5.2.0
edge-tts>=6.1.0
groq>=0.9.0
psutil>=5.9.0
pyautogui>=0.9.54
pyaudio>=0.2.14
pygame>=2.5.0
pyperclip>=1.8.2
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
speechrecognition>=3.10.0
spotipy>=2.23.0
vosk>=0.3.45
